requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
babel==2.7.0              # via prices
boto3==1.9.169            # via django-amazon-ses
botocore==1.12.169        # via boto3, s3transfer
certifi==2019.3.9         # via requests, sentry-sdk
chardet==3.0.4            # via requests
defusedxml==0.6.0         # via python3-openid
django-allauth==0.39.1
django-amazon-ses==2.1.1
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django==2.2.1
djangorestframework==3.9.3
docutils==0.14            # via botocore
gevent==1.4.0             # via gunicorn
google-measurement-protocol==1.0.0
greenlet==0.4.15          # via gevent
gunicorn[gevent]==19.9.0
idna==2.8                 # via requests, tldextract
jmespath==0.9.4           # via boto3, botocore
oauthlib==3.0.1           # via requests-oauthlib
prices==1.0.0             # via google-measurement-protocol
python-dateutil==2.8.0    # via botocore
python3-openid==3.1.0     # via django-allauth
pytz==2019.1              # via babel, django
requests-file==1.4.3      # via tldextract
requests-oauthlib==1.2.0  # via django-allauth
requests==2.21.0          # via django-allauth, google-measurement-protocol, requests-file, requests-oauthlib, tldextract
s3transfer==0.2.1         # via boto3
sentry-sdk==0.9.0
six==1.12.0               # via python-dateutil, requests-file
sqlparse==0.3.0           # via django
tldextract==2.2.1
typing==3.6.6             # via google-measurement-protocol, prices
urllib3==1.24.3           # via botocore, requests, sentry-sdk

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via tldextract
